glnx-errors.h: add glnx_null_throw[_*] variants
These are equivalent to the non-null throw, except that the returned
value is a NULL pointer. They can be used in functions where one wants
to return a pointer. E.g.:

	GKeyFile *foo(GError **error) {
		return glnx_null_throw (error, "foobar");
	}

They are of course not semantically equivalent:

	void foo(void) {
		glnx_throw (error, "foo");
		glnx_null_throw (error, "bar");
	}

The first throw is an expression statement, while the second is a pure
expression. Thus, the latter will cause a -Wunused-value warning. Though
in such situations, there's no point in using the null_ variant over the
regular function call in the first place.

I made the 'args...' absorb the fmt argument as well so that callers can
still use it without always having to specify at least one additional
variadic argument. I had to check to be sure that the expansion is all
done by the preprocessor, so we don't need to worry about stack
intricacies.